ci: Fixing workflow runs for renovate and molecule on forked PRs
Fixes #188

The molecule and renovate workflows make use of secrets. When a PR
is created which involves changes to either the molecule or the
renovate workflow, it will fail, as the secrets are inaccessible
to the PR.

This commit introduces a new action to check if the person that
triggered the PR has write permissions on the repository. If not
the workflow will not run.

If a contributor will re-run the failed workflow, it will use the
permissions of that contributor and check out the PRs HEAD SHA
and run against that.